M_HU_Unique_Attributes https://github.com/metasfresh/metasfresh/issues/13850
