+++ backend/de.metas.business/src/main/java/de/metas/uom/IUOMDAO.java
@@ -76,6 +76,8 @@ public interface IUOMDAO extends ISingletonService
 	 */
 	boolean isUOMForTUs(UomId uomId);
 
+	boolean isUOMEach(@NonNull UomId uomId);
+
 	@NonNull
 	UOMType getUOMTypeById(UomId uomId);
 }
