+++ backend/de.metas.business/src/main/java/de/metas/uom/impl/UOMDAO.java
@@ -196,6 +196,13 @@ public class UOMDAO implements IUOMDAO
 	}
 
 	@Override
+	public boolean isUOMEach(@NonNull final UomId uomId)
+	{
+		final X12DE355 x12de355 = getX12DE355ById(uomId);
+		return X12DE355.EACH.equals(x12de355);
+	}
+
+	@Override
 	public @NonNull UOMType getUOMTypeById(@NonNull final UomId uomId)
 	{
 		final I_C_UOM uom = getById(uomId);
